refactor: rename Toc with DocsToc and enhance code readability
