--- full-kit/src/app/(unlocalized)/docs/_components/toggle-mobile-sidebar.tsx
@@ -1,24 +1,0 @@
-"use client"
-
-import { PanelLeft } from "lucide-react"
-
-import { Button } from "@/components/ui/button"
-import { useSidebar } from "@/components/ui/sidebar"
-
-export function ToggleMobileSidebar() {
-  const { isMobile, openMobile, setOpenMobile } = useSidebar()
-
-  if (isMobile) {
-    return (
-      <Button
-        data-sidebar="trigger"
-        variant="ghost"
-        size="icon"
-        onClick={() => setOpenMobile(!openMobile)}
-      >
-        <PanelLeft className="h-4 w-4" />
-        <span className="sr-only">Toggle Sidebar</span>
-      </Button>
-    )
-  }
-}
